migrate load logic to its own file
test plan: ensure no load errors on render in initiatives or cred
explorer when building and browsing the dev site
